Avoid copying entire cache on each optimistic read.
Previously, the InMemoryCache maintained an array of recorded optimistic
updates, which it would reapply to an entirely new ObjectCache whenever
performing any single optimistic read.

This reapplication was wasteful, but the real performance killer was
calling this.extract(true) each time, which copied the entire underlying
(non-optimistic) cache, just to create a throw-away ObjectCache for the
duration of a single optimistic read. Worse still, this.extract(true) was
also called in recordOptimisticTransaction, just to create a throw-away
RecordingCache object.

Instead of creating temporary copies of the entire cache, InMemoryCache
now maintains a linked list of OptimisticCacheLayer objects, which extend
ObjectCache and implement the NormalizedCache interface, but cleverly
delegate to a parent cache for any missing properties. This delegation
happens simply by calling the .get(dataId) method of the parent cache, so
there is no need to extract/copy the parent cache.

When no optimistic data are currently active, cache.optimisticData ===
cache.data, which means there are no additional layers on top of the
actual data. Lookup time is proportional to the number of
OptimisticCacheLayer objects in the linked list, so it's best if that list
remains reasonably short, but at least that's something the application
developer can control.

Calling removeOptimistic(id) removes all OptimisticCacheLayer objects with
the given id, without affecting any other layers or the underlying data.